gerar XML e JSON

+++ TheatricalPlayersRefactoringKata/Entities/Statement.cs
@@ -7,6 +7,6 @@ public class Statement
     public int Id { get; set; }
     public string Customer { get; set; }
     public decimal TotalAmountOwed { get; set; }
-    public int TotalEarnedCredits { get; set; }
+    public decimal TotalEarnedCredits { get; set; }
     public List<StatementItem> Items { get; set; }
 }
